feat: cambio de archivo .py para el router.get de sintetizar_ri
